Acceptance: Get rid of `defer` in `ECS` tests
Replace `defer` for quotas with `quotas.Book...` calls
