+++ opentelekomcloud/acceptance/ecs/data_source_opentelekomcloud_compute_flavor_v2_test.go
@@ -34,6 +34,8 @@ func TestAccComputeV2FlavorDataSource_basic(t *testing.T) {
 }
 
 func TestAccComputeV2FlavorDataSource_testQueries(t *testing.T) {
+	t.Parallel()
+
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
 		ProviderFactories: common.TestAccProviderFactories,
