--- opentelekomcloud/acceptance/ecs/import_opentelekomcloud_compute_floatingip_v2_test.go
@@ -1,33 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccComputeV2FloatingIP_importBasic(t *testing.T) {
-	t.Parallel()
-	th.AssertNoErr(t, quotas.FloatingIP.Acquire())
-	defer quotas.FloatingIP.Release()
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckComputeV2FloatingIPDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccComputeV2FloatingIPBasic,
-			},
-			{
-				ResourceName:      resourceFloatingIpName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
