--- opentelekomcloud/acceptance/ecs/import_opentelekomcloud_compute_instance_v2_test.go
@@ -1,62 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/env"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccComputeV2InstanceImportBasic(t *testing.T) {
-	qts := serverQuotas(4, env.OsFlavorID)
-	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      TestAccCheckComputeV2InstanceDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccComputeV2InstanceBasic,
-			},
-			{
-				ResourceName:      resourceInstanceV2Name,
-				ImportState:       true,
-				ImportStateVerify: true,
-				ImportStateVerifyIgnore: []string{
-					"stop_before_destroy",
-					"force_delete",
-				},
-			},
-		},
-	})
-}
-
-func TestAccComputeV2InstanceImportBootFromVolumeImage(t *testing.T) {
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      TestAccCheckComputeV2InstanceDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccComputeV2InstanceBootFromVolumeImage,
-			},
-			{
-				ResourceName:      resourceInstanceV2Name,
-				ImportState:       true,
-				ImportStateVerify: true,
-				ImportStateVerifyIgnore: []string{
-					"stop_before_destroy",
-					"force_delete",
-				},
-			},
-		},
-	})
-}
