--- opentelekomcloud/acceptance/ecs/import_opentelekomcloud_compute_servergroup_v2_test.go
@@ -1,31 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccComputeV2ServerGroup_importBasic(t *testing.T) {
-	resourceName := "opentelekomcloud_compute_servergroup_v2.sg_1"
-	t.Parallel()
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckComputeV2ServerGroupDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccComputeV2ServerGroupBasic,
-			},
-
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
