--- opentelekomcloud/acceptance/ecs/import_opentelekomcloud_compute_volume_attach_v2_test.go
@@ -1,36 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccComputeV2VolumeAttach_importBasic(t *testing.T) {
-	t.Parallel()
-	qts := serverQuotas(1+4, "s2.medium.1")
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckComputeV2VolumeAttachDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccComputeV2VolumeAttachBasic,
-			},
-
-			{
-				ResourceName:      resourceVolumeAttach,
-				ImportState:       true,
-				ImportStateVerify: true,
-			},
-		},
-	})
-}
