--- opentelekomcloud/acceptance/ecs/import_opentelekomcloud_ecs_instance_v1_test.go
@@ -1,40 +1,0 @@
-package acceptance
-
-import (
-	"testing"
-	"time"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccEcsV1Instance_importBasic(t *testing.T) {
-	t.Parallel()
-	qts := serverQuotas(10+4, "s2.medium.1")
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { common.TestAccPreCheck(t) },
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckEcsV1InstanceDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccEcsV1InstanceBasic,
-			},
-
-			{
-				ResourceName:      resourceInstanceV1Name,
-				ImportState:       true,
-				ImportStateVerify: true,
-				ImportStateVerifyIgnore: []string{
-					"password",
-					"delete_disks_on_termination",
-				},
-			},
-		},
-	})
-}
