+++ opentelekomcloud/acceptance/ecs/resource_opentelekomcloud_compute_floatingip_v2_test.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
 
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/compute/v2/extensions/floatingips"
@@ -21,8 +20,8 @@ const resourceFloatingIpName = "opentelekomcloud_compute_floatingip_v2.fip_1"
 func TestAccComputeV2FloatingIP_basic(t *testing.T) {
 	var fip floatingips.FloatingIP
 
-	th.AssertNoErr(t, quotas.FloatingIP.Acquire())
-	defer quotas.FloatingIP.Release()
+	t.Parallel()
+	quotas.BookOne(t, quotas.FloatingIP)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -34,6 +33,27 @@ func TestAccComputeV2FloatingIP_basic(t *testing.T) {
 				Check: resource.ComposeTestCheckFunc(
 					testAccCheckComputeV2FloatingIPExists(resourceFloatingIpName, &fip),
 				),
+			},
+		},
+	})
+}
+
+func TestAccComputeV2FloatingIP_importBasic(t *testing.T) {
+	t.Parallel()
+	quotas.BookOne(t, quotas.FloatingIP)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { common.TestAccPreCheck(t) },
+		ProviderFactories: common.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckComputeV2FloatingIPDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccComputeV2FloatingIPBasic,
+			},
+			{
+				ResourceName:      resourceFloatingIpName,
+				ImportState:       true,
+				ImportStateVerify: true,
 			},
 		},
 	})
