+++ opentelekomcloud/acceptance/ecs/resource_opentelekomcloud_compute_instance_v2_test.go
@@ -4,11 +4,9 @@ import (
 	"fmt"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	th "github.com/opentelekomcloud/gophertelekomcloud/testhelper"
 	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common/quotas"
 
 	"github.com/opentelekomcloud/gophertelekomcloud/openstack/compute/v2/extensions/secgroups"
@@ -27,8 +25,7 @@ func TestAccComputeV2Instance_basic(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -58,6 +55,32 @@ func TestAccComputeV2Instance_basic(t *testing.T) {
 	})
 }
 
+func TestAccComputeV2Instance_importBasic(t *testing.T) {
+	t.Parallel()
+	qts := serverQuotas(4, env.OsFlavorID)
+	quotas.BookMany(t, qts)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { common.TestAccPreCheck(t) },
+		ProviderFactories: common.TestAccProviderFactories,
+		CheckDestroy:      TestAccCheckComputeV2InstanceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccComputeV2InstanceBasic,
+			},
+			{
+				ResourceName:      resourceInstanceV2Name,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateVerifyIgnore: []string{
+					"stop_before_destroy",
+					"force_delete",
+				},
+			},
+		},
+	})
+}
+
 func TestAccComputeV2Instance_multiSecgroup(t *testing.T) {
 	var instance servers.Server
 	var secGroup1, secGroup2 secgroups.SecurityGroup
@@ -65,8 +88,7 @@ func TestAccComputeV2Instance_multiSecgroup(t *testing.T) {
 	secGroupName2 := "opentelekomcloud_compute_secgroup_v2.secgroup_2"
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -97,8 +119,7 @@ func TestAccComputeV2Instance_bootFromImage(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(50, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -120,8 +141,7 @@ func TestAccComputeV2Instance_bootFromVolume(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(50, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -138,12 +158,37 @@ func TestAccComputeV2Instance_bootFromVolume(t *testing.T) {
 	})
 }
 
+func TestAccComputeV2Instance_importBootFromVolumeImage(t *testing.T) {
+	t.Parallel()
+	qts := serverQuotas(4, env.OsFlavorID)
+	quotas.BookMany(t, qts)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { common.TestAccPreCheck(t) },
+		ProviderFactories: common.TestAccProviderFactories,
+		CheckDestroy:      TestAccCheckComputeV2InstanceDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccComputeV2InstanceBootFromVolumeImage,
+			},
+			{
+				ResourceName:      resourceInstanceV2Name,
+				ImportState:       true,
+				ImportStateVerify: true,
+				ImportStateVerifyIgnore: []string{
+					"stop_before_destroy",
+					"force_delete",
+				},
+			},
+		},
+	})
+}
+
 func TestAccComputeV2Instance_changeFixedIP(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -164,8 +209,7 @@ func TestAccComputeV2Instance_bootFromVolumeVolume(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(50, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -187,8 +231,7 @@ func TestAccComputeV2Instance_stopBeforeDestroy(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -209,8 +252,7 @@ func TestAccComputeV2Instance_metadata(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -246,8 +288,7 @@ func TestAccComputeV2Instance_timeout(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -268,8 +309,7 @@ func TestAccComputeV2Instance_autoRecovery(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -298,8 +338,7 @@ func TestAccComputeV2Instance_crazyNICs(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck:          func() { common.TestAccPreCheck(t) },
@@ -320,8 +359,7 @@ func TestAccComputeV2Instance_initialStateActive(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
@@ -362,8 +400,7 @@ func TestAccComputeV2Instance_initialStateShutoff(t *testing.T) {
 	var instance servers.Server
 	qts := serverQuotas(4, env.OsFlavorID)
 	t.Parallel()
-	th.AssertNoErr(t, quotas.AcquireMultipleQuotas(qts, 5*time.Second))
-	defer quotas.ReleaseMultipleQuotas(qts)
+	quotas.BookMany(t, qts)
 
 	resource.Test(t, resource.TestCase{
 		PreCheck: func() {
