+++ opentelekomcloud/acceptance/ecs/resource_opentelekomcloud_compute_secgroup_v2_test.go
@@ -35,6 +35,26 @@ func TestAccComputeV2SecGroup_basic(t *testing.T) {
 	})
 }
 
+func TestAccComputeV2SecGroup_importBasic(t *testing.T) {
+	t.Parallel()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:          func() { common.TestAccPreCheck(t) },
+		ProviderFactories: common.TestAccProviderFactories,
+		CheckDestroy:      testAccCheckComputeV2SecGroupDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccComputeV2SecGroupBasicOrig,
+			},
+			{
+				ResourceName:      resourceSecGroupName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
 func TestAccComputeV2SecGroup_update(t *testing.T) {
 	var secGroup secgroups.SecurityGroup
 	t.Parallel()
